pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metahan"
version = "0.1.0"
description = "Metasurface layout generation and GDSII export toolkit."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gdstk",
    "numpy",
    "pyyaml",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
]

[project.scripts]
metahan = "metahan.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
